@@ -2,7 +2,7 @@ source 'https://rubygems.org'
 ruby '2.4.4'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.10'
+gem 'rails', '5.2.3'
 
 # Use PG as the database for Active Record
 gem 'pg', '~> 0.21.0'
@@ -23,7 +23,7 @@ gem 'typhoeus'
 gem 'diffy'
 
 # XML
-gem 'nokogiri', '~> 1.8.5'
+gem 'nokogiri'
 
 # CORS
 gem 'rack-cors', :require => 'rack/cors'
@@ -32,7 +32,7 @@ gem 'rack-cors', :require => 'rack/cors'
 gem 'uglifier', '>= 1.3.0'
 
 # Use CoffeeScript for .coffee assets and views
-gem 'coffee-rails', '~> 4.1.0'
+gem 'coffee-rails', '~> 5.0.0'
 
 # See https://github.com/rails/execjs#readme for more supported runtimes
 # gem 'therubyracer', platforms: :ruby
@@ -41,9 +41,9 @@ gem 'jquery-rails'
 gem "jquery-validation-rails"
 
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 2.0'
+gem 'jbuilder', '>= 2.0'
 # bundle exec rake doc:rails generates the API under doc/api.
-gem 'sdoc', '~> 0.4.0', group: :doc
+gem 'sdoc', '>= 0.4.0', group: :doc
 
 # D3 gem
 gem "d3-rails", "3.4.11"
@@ -90,7 +90,6 @@ group :development do
   gem 'web-console', '~> 2.0'
   gem 'better_errors'
   gem 'hub', :require=>nil
-  gem 'quiet_assets'
   gem 'rails_layout'
 end
 
