@@ -1,4 +1,4 @@
-class Role < ActiveRecord::Base
+class Role < ApplicationRecord
 
   has_and_belongs_to_many :users, :join_table => :users_roles
 
